+++ NPCVehicleSync/Client/serverEvents.js
@@ -1,0 +1,8 @@
+import * as alt from 'alt'
+import { NPCVehHandler } from './npcVehicleHandler.js';
+
+alt.log("[NPCVehicleHandler] Loaded ServerEvents")
+
+alt.onServer("NPCVehicle:NetOwner", (vehicle, state, transfer = undefined) => {
+    NPCVehHandler.UpdateNetOwner(vehicle, state, transfer)
+})
